+++ packages/adapter-react/src/files/variant.js
@@ -1,7 +1,0 @@
-module.exports = (upcasedComponentName) => `
-import React from 'react'
-import ${upcasedComponentName} from '../${upcasedComponentName}.jsx'
-
-export default props => (
-  <${upcasedComponentName} {...props} />
-)`