+++ packages/adapter-react/test/fixtures/elements/Atom/variants/Atom.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import Atom from '../Atom'
-
-export default props => (
-  <Atom {...props} />
-)