+++ packages/adapter-react/test/fixtures/invalid-template.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Template (props) {
-  return (
-    <p>
-      {props.}
-    </p>
-  )
-}