+++ packages/adapter-react/test/fixtures/modules/Molecule/Molecule.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import Atom from '../../elements/Atom'
-
-export default props => (
-  <div className='molecule'>
-    <Atom />
-  </div>
-)