+++ packages/adapter-react/test/fixtures/modules/Molecule/variants/Molecule.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import Molecule from '../Molecule'
-
-export default props => (
-  <Molecule {...props} />
-)