+++ packages/adapter-vue/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  root: false,
-  parserOptions: {
-    parser: 'babel-eslint'
-  },
-  extends: [
-    'plugin:vue/recommended'
-  ]
-}