+++ packages/adapter-vue/test/fixtures/props.js
@@ -1,6 +1,0 @@
-export default {
-  myData: {
-    type: Number,
-    required: true
-  }
-}