+++ packages/adapter-vue/test/fixtures/template-export.js
@@ -1,9 +1,0 @@
-import props from './props'
-
-module.exports = {
-  name: 'TestTemplate',
-
-  props,
-
-  template: '<p>{{ myData }}</p>'
-}