+++ packages/adapter-vue/test/fixtures/template-register-no-export.js
@@ -1,8 +1,0 @@
-import Vue from 'vue'
-import props from './props'
-
-Vue.component('test-template', {
-  props,
-
-  template: '<p>{{ myData }}</p>'
-})