+++ test/project/lib/react-adapter-options.js
@@ -1,8 +1,0 @@
-module.exports = {
-  babel: {
-    presets: [
-      '@babel/preset-env',
-      '@babel/preset-react'
-    ]
-  }
-}